+++ src/main/java/commom/graph/Edge.java
@@ -3,9 +3,9 @@ package main.java.commom.graph;
 public class Edge {
     public final Node nodeFrom;
     public final Node nodeTo;
-    public final int weight;
+    public final double weight;
 
-    public Edge(Node nodeFrom, Node nodeTo, int weight){
+    public Edge(Node nodeFrom, Node nodeTo, double weight){
         this.nodeFrom = nodeFrom;
         this.nodeTo = nodeTo;
         this.weight = weight;
